@@ -1,14 +1,13 @@
 import React from "react";
-import ReactDOM from "react-dom";
-import CssBaseline from '@mui/material/CssBaseline';
+import { createRoot } from "react-dom/client";
+import CssBaseline from "@mui/material/CssBaseline";
 import "./index.css";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
 
-ReactDOM.render(
-    <><CssBaseline /><App /></>,
-  document.getElementById("root")
-);
+const container = document.getElementById("root");
+const root = createRoot(container);
+root.render(<><CssBaseline /><App /></>);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
