@@ -1,11 +1,11 @@
-import React from 'react';
-import SelectDate from './SelectDate';
-import {AppContextProvider} from '../../context/AppContext';
-import {LocalizationProvider} from "@mui/lab";
-import AdapterDateFns from "@mui/lab/AdapterDateFns";
+import React from "react";
+import SelectDate from "./SelectDate";
+import {AppContextProvider} from "../../context/AppContext";
+import { LocalizationProvider } from "@mui/x-date-pickers";
+import { AdapterDateFns } from "@mui/x-date-pickers/AdapterDateFns";
 
 export default {
-    title: 'Check-Ins/SelectDate',
+    title: "Check-Ins/SelectDate",
     component: SelectDate,
     args: {
       changeQuery: (data1, data2) => {}
