@@ -2,7 +2,10 @@ import React from "react";
 import FeedbackRecipientSelector from "./FeedbackRecipientSelector";
 import {AppContextProvider} from "../../context/AppContext";
 import {BrowserRouter} from "react-router-dom";
+import {jest} from '@jest/globals';
 
-it("renders the component", () => {
-  snapshot(<BrowserRouter><AppContextProvider><FeedbackRecipientSelector /></AppContextProvider></BrowserRouter>);
+describe("FeedbackRecipientSelector", () => {
+  it("renders the component", () => {
+    snapshot(<BrowserRouter><AppContextProvider><FeedbackRecipientSelector changeQuery={jest.fn()} fromQuery={[]} forQuery="" /></AppContextProvider></BrowserRouter>);
+  });
 });
