@@ -1,18 +1,20 @@
 import React from "react";
+import ReactDOM from "react-dom";
+import {AppContextProvider} from "../../context/AppContext";
 import SkillCategoryNewDialog from "./SkillCategoryNewDialog";
-import EnzymeToJson from 'enzyme-to-json';
-import { mount } from 'enzyme';
 
-
-it("renders correctly", () => {
-  const dialog = mount(
-    <SkillCategoryNewDialog
-      isOpen={true}
-      onClose={jest.fn()}
-      onConfirm={jest.fn()}
-    />
-  );
-  expect(EnzymeToJson(dialog)).toMatchSnapshot({});
+describe("SkillCategoryNewDialog", () => {
+  it("renders correctly", () => {
+    snapshot(
+      <AppContextProvider>
+        <SkillCategoryNewDialog
+          isOpen={true}
+          onClose={jest.fn()}
+          onConfirm={jest.fn()}
+        />
+      </AppContextProvider>
+    );
+  });
 });
 
 
