@@ -2,8 +2,25 @@ import React from "react";
 import FeedbackResponseCard from "./FeedbackResponseCard";
 import {AppContextProvider} from "../../../context/AppContext";
 
+import {jest} from "@jest/globals";
+
+jest.mock("@mui/material/Slider", () => (props: any) => {
+  const { onChange, 'data-testid': testId, ...rest } = props;
+
+  return (
+    <input
+      data-testid={testId}
+      type="range"
+      onChange={event => {
+        onChange(null, parseInt(event.target.value, 10));
+      }}
+      {...rest}
+    />
+  );
+});
+
 it("renders correctly for text responses", () => {
-  shallowSnapshot(
+  snapshot(
     <AppContextProvider>
       <FeedbackResponseCard
         responderId="01b7d769-9fa2-43ff-95c7-f3b950a27bf9"
@@ -15,7 +32,7 @@ it("renders correctly for text responses", () => {
 });
 
 it("renders correctly for radio button responses", () => {
-  shallowSnapshot(
+  snapshot(
     <AppContextProvider>
       <FeedbackResponseCard
         responderId="01b7d769-9fa2-43ff-95c7-f3b950a27bf9"
@@ -27,7 +44,7 @@ it("renders correctly for radio button responses", () => {
 });
 
 it("renders correctly for slider responses", () => {
-  shallowSnapshot(
+  snapshot(
     <AppContextProvider>
       <FeedbackResponseCard
         responderId="01b7d769-9fa2-43ff-95c7-f3b950a27bf9"
